attempted to see if fixing the read bug fixed extraction too unfortunately not also fix up types and add set methods on pokemon

+++ Pokemon-XD-Common/PokemonDefinitions/Evolution.cs
@@ -8,9 +8,9 @@ namespace XDCommon.PokemonDefinitions
     {
         public EvolutionMethods EvolutionMethod { get; }
         public EvolutionConditionType EvolutionCondition { get; }
-        public int EvolvesInto { get; }
+        public ushort EvolvesInto { get; }
 
-        public Evolution(int method, int condition, int evolvesInto)
+        public Evolution(byte method, ushort condition, ushort evolvesInto)
         {
             EvolutionMethod = (EvolutionMethods)method;
             EvolutionCondition = (EvolutionConditionType)condition;
